backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
python-multipart==0.0.9
PyNaCl==1.5.0

numpy==1.26.4
matplotlib==3.9.2
pytest==8.3.3
httpx==0.27.2

